+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLCredentialQuerySerializer.kt
@@ -7,15 +7,17 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 
-object DCQLCredentialQuerySerializer : JsonContentPolymorphicSerializer<DCQLCredentialQuery>(DCQLCredentialQuery::class) {
+object DCQLCredentialQuerySerializer :
+    JsonContentPolymorphicSerializer<DCQLCredentialQuery>(DCQLCredentialQuery::class) {
     override fun selectDeserializer(element: JsonElement): DeserializationStrategy<DCQLCredentialQuery> {
         val parameters = element.jsonObject
-        val credentialFormatIdentifier = parameters[DCQLCredentialQuery.SerialNames.FORMAT]?.jsonPrimitive?.content?.let {
-            CredentialFormatEnum.parse(it)
-        }
+        val credentialFormatIdentifier =
+            parameters[DCQLCredentialQuery.SerialNames.FORMAT]?.jsonPrimitive?.content?.let {
+                CredentialFormatEnum.parse(it)
+            }
         return when {
             credentialFormatIdentifier == CredentialFormatEnum.MSO_MDOC -> DCQLIsoMdocCredentialQuery.serializer()
-            credentialFormatIdentifier?.coerceDeprecations() == CredentialFormatEnum.DC_SD_JWT -> DCQLSdJwtCredentialQuery.serializer()
+            credentialFormatIdentifier == CredentialFormatEnum.DC_SD_JWT -> DCQLSdJwtCredentialQuery.serializer()
             else -> DCQLCredentialQueryInstance.serializer()
         }
     }