+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/PresentationDataClasses.kt
@@ -13,10 +13,7 @@ import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.signum.indispensable.josef.JwsSigned
-import at.asitplus.wallet.lib.data.Base64URLTransactionDataSerializer
-import at.asitplus.wallet.lib.data.SdJwtConstants
 import at.asitplus.wallet.lib.data.VerifiablePresentationJws
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.jws.SdJwtSigned
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.Serializable
@@ -25,6 +22,7 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonPrimitive
 import kotlinx.serialization.json.buildJsonArray
 import kotlinx.serialization.json.buildJsonObject
+import kotlinx.serialization.json.putJsonArray
 
 /**
  * Input to create a verifiable presentation of credentials, i.e. contains input required to fill fields in the VP,
@@ -38,15 +36,11 @@ data class PresentationRequestParameters(
     val nonce: String,
     val audience: String,
     val transactionData: List<TransactionDataBase64Url>? = null,
-    @Deprecated("Use calcIsoDeviceSignaturePlain instead")
-    val calcIsoDeviceSignature: (suspend (docType: String, deviceNameSpaceBytes: ByteStringWrapper<DeviceNameSpaces>) -> Pair<CoseSigned<ByteArray>, String?>?) =
-        { _, _ -> null },
     /**
      * Handle calculating device signature for ISO mDocs, as this depends on the transport protocol
      * (OpenID4VP with ISO/IEC 18013-7)
      */
-    val calcIsoDeviceSignaturePlain: (suspend (input: IsoDeviceSignatureInput) -> CoseSigned<ByteArray>?) =
-        { calcIsoDeviceSignature(it.docType, it.deviceNameSpaceBytes)?.first },
+    val calcIsoDeviceSignaturePlain: (suspend (input: IsoDeviceSignatureInput) -> CoseSigned<ByteArray>?) = { null },
     /**
      * mdocGeneratedNonce to be used for the presentation and [calcIsoDeviceSignaturePlain]
      * (OpenID4VP with ISO/IEC 18013-7)
@@ -62,12 +56,6 @@ data class PresentationRequestParameters(
      * For convenience, we always select the first if the set is non-empty
      */
     val transactionDataHashesAlgorithm: Digest? = getCommonHashesAlgorithms(transactionData)?.first().toDigest()
-
-    @Deprecated("No longer necessary. Will be removed")
-    enum class Flow {
-        OID4VP,
-        UC5
-    }
 }
 
 data class IsoDeviceSignatureInput(
@@ -85,8 +73,11 @@ sealed interface PresentationResponseParameters {
     ) : PresentationResponseParameters {
         override val vpToken
             get() = buildJsonObject {
+                // To be reconsidered when supporting [DCQLCredentialQueryInstance.multiple]
                 verifiablePresentations.entries.forEach {
-                    put(it.key.string, it.value.toJsonPrimitive())
+                    putJsonArray(it.key.string) {
+                        add(it.value.toJsonPrimitive())
+                    }
                 }
             }